requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==2.0.2
django-cors-headers==2.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
idna==2.6
mysqlclient==1.3.12
Pillow==5.0.0
PyJWT==1.5.3
python-http-client==3.0.0
pytz==2018.3
requests==2.18.4
sendgrid==3.6.5
sendgrid-django==4.2.0
stripe==1.77.2
urllib3==1.22
